ordre historie view lagt til. Oppdatert user kontroller og DBProduct til å støtte dette. Endre toObject metoden i DBProduct til static. Endret int til double i OrderModel. DBOrder lagt til. Måtte endre LoginView for å få kjørt.

+++ Kaffeplaneten/Kaffeplaneten/Models/OrderModel.cs
@@ -10,6 +10,6 @@ namespace Kaffeplaneten.Models
         public int orderNr { get; set; }
         public int customerID { get; set; }
         public List<ProductModel> products { get; set; }//liste over alle produkter i orderen. Legg inn dublikater hvis flere av samme produkt er bestilt.
-        public int total { get; set; }//total pris for orderen
+        public double total { get; set; }//total pris for orderen
     }
 }
